Upgrade document check descriptor test
